@@ -5,6 +5,10 @@ import sys
 
 
 def checkBAMtag(bam, name, tag):
+    """
+    Check whether a given tag is present in the BAM file.
+    """
+
     bctag = [read.has_tag(tag) for read in bam.head(1000)]
     if not any(bctag):
         sys.stderr.write("WARNING: Input file {} seems to lack the tag {}. Output might be empty. \n".format(name, tag))
@@ -19,18 +23,25 @@ def checkMotifs(read, chrom, genome, readMotif, refMotif):
     like this below.
 
     Forwards aligned read: read has 'A', upstream has T
-    R1 ........A------->
-    ----------TA------------\ Ref (+)
 
-    Rev aligned read: read has 'T', downstream has A
+        R1 ........A------->
+        ----------TA------------\ Ref (+)
 
-    <-------T....... R1
-    --------TA------------\ Ref (+)
+        Rev aligned read: read has 'T', downstream has A
+
+        <-------T....... R1
+        --------TA------------\ Ref (+)
 
     This function can look for any arbitrary motif in read and corresponding genome, but in the
     same orientation as described above.
 
-    :return: bool
+    Returns
+    -------
+    bool
+        True if the motif is present in the read and reference genome.
+
+    Examples
+    --------
 
     >>> import pysam
     >>> import os
@@ -113,7 +124,10 @@ def checkAlignedFraction(read, lowFilter):
     the given threshold. Aligned fraction includes the max allowed mismatches tolerated by
     the aligner, and excludes InDels and Clippings.
 
-    Return: Bool
+    Returns
+    -------
+    bool
+        True if the fraction of read length aligns to the reference.
     """
     cig = read.cigartuples
     tot = read.infer_read_length()
