@@ -1,1 +1,1 @@
-__version__ = "0.5.1"
+__version__ = "0.5.2"
