@@ -75,13 +75,14 @@ def parseArguments(args=None):
     parser = argparse.ArgumentParser(
         parents=[io_args, bam_args, read_args, filter_args, get_args(), other_args],
         formatter_class=argparse.ArgumentDefaultsHelpFormatter,
-        description="This tool samples regions in the genome from BAM files "
-        "and compares the cumulative read coverages for each cell on those regions. "
-        "to a synthetic cell with poisson distributed reads using Jansson Shannon Distance. "
-        "Cells with high enrichment of signals show a higher JSD compared to cells whose signal "
-        "is homogenously distrubuted.",
+        description="""
+``scJSD`` samples regions in the genome from BAM files and compares the cumulative read coverages
+for each cell on those regions to a synthetic cell with poisson distributed reads using the Jensen-Shannon
+distance. Cells with high enrichment of signals show a higher JSD compared to cells whose signal is
+homogeneously distributed.
+""",
         conflict_handler="resolve",
-        usage="An example usage is: plotFingerprint -b treatment.bam control.bam " "-plot fingerprint.png",
+        usage="plotFingerprint -b treatment.bam control.bam " "-plot fingerprint.png",
         add_help=False,
     )
 
